requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1.3
django-rest-swagger==2.2.0
djangorestframework==3.9.0
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
openapi-codec==1.3.2
opencv-python==3.4.3.18
Pillow==5.3.0
psycopg2==2.7.7
pytesseract==0.2.6
pytz==2018.7
PyYAML==3.13
requests==2.21.0
ruamel.yaml==0.15.85
simplejson==3.16.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
kraken